requirements.txt
paho-mqtt==2.1.0
#pydantic==2.9.2
#pydantic_core==2.23.4
#python==3.12.3
#fastapi==0.112.1
#node==v20.16.0
#npm==10.8.3
#react==18.3.1
vite==1.5.2
#alembic==1.13.3
#mosqitto==2.0.18
#sqlite==3.37.2
#alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.1
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
ecdsa==0.19.0
fastapi==0.115.5
h11==0.14.0
idna==3.10
jwt==1.3.1
Mako==1.3.6
MarkupSafe==3.0.2
mqtt==0.0.1
paho-mqtt==2.1.0
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.0
pydantic_core==2.27.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1